Add Double precision

+++ src/main/java/fr/pollux/katarpn/Operation.java
@@ -2,7 +2,7 @@ package fr.pollux.katarpn;
 
 public class Operation
 {
-    public static int addition(int num1, int num2)
+    public static double addition(double num1, double num2)
     {
         return num1 + num2;
     }
